Juego favorito para el usuario añadido. Refs #11

+++ frontend/app/src/main/java/com/example/playtracker/data/api/GameApi.kt
@@ -14,5 +14,5 @@ interface GameApi {
     suspend fun getPopularGames(@Query("page") page: Int = 2): List<Game>
 
     @GET("rawg/games/{id}")
-    suspend fun getGameDetails(@Path("id") gameId: Int): GameDetail
+    suspend fun getGameDetails(@Path("id") gameId: Long): GameDetail
 }
